@@ -26,33 +26,33 @@
     "devDependencies": {
         "@angular-material-extensions/select-country": "^8.1.0",
         "@angular-eslint/eslint-plugin-template": "15.2.1",
-        "@angular/platform-browser-dynamic": "~15.2.3",
-        "@angular/material-moment-adapter": "^15.2.3",
+        "@angular/platform-browser-dynamic": "~15.2.4",
+        "@angular/material-moment-adapter": "^15.2.4",
         "@angular-devkit/build-webpack": "^0.1502.4",
         "@angular-eslint/template-parser": "15.2.1",
         "@angular-devkit/build-angular": "^15.2.4",
         "@angular-eslint/eslint-plugin": "15.2.1",
         "@angular-eslint/schematics": "15.2.1",
-        "@angular/platform-browser": "~15.2.3",
+        "@angular/platform-browser": "~15.2.4",
         "@angular-eslint/builder": "15.2.1",
-        "@angular/compiler-cli": "~15.2.3",
-        "@angular/animations": "~15.2.3",
+        "@angular/compiler-cli": "~15.2.4",
+        "@angular/animations": "~15.2.4",
         "@schematics/angular": "^15.2.4",
-        "@angular/compiler": "~15.2.3",
-        "@angular/material": "^15.2.3",
-        "@angular/router": "~15.2.3",
-        "@angular/common": "~15.2.3",
-        "@angular/forms": "~15.2.3",
-        "@angular/core": "~15.2.3",
-        "@angular/cdk": "^15.2.3",
+        "@angular/compiler": "~15.2.4",
+        "@angular/material": "^15.2.4",
+        "@angular/router": "~15.2.4",
+        "@angular/common": "~15.2.4",
+        "@angular/forms": "~15.2.4",
+        "@angular/core": "~15.2.4",
+        "@angular/cdk": "^15.2.4",
         "@angular/cli": "~15.2.4",
         "@biesbjerg/ngx-translate-extract-marker": "^1.0.0",
         "@bartholomej/ngx-translate-extract": "^8.0.2",
-        "@typescript-eslint/eslint-plugin": "5.56.0",
-        "@fortawesome/fontawesome-free": "^6.3.0",
+        "@typescript-eslint/eslint-plugin": "5.57.0",
+        "@fortawesome/fontawesome-free": "^6.4.0",
         "eslint-plugin-prefer-arrow": "latest",
         "@ngx-translate/http-loader": "^7.0.0",
-        "@typescript-eslint/parser": "5.56.0",
+        "@typescript-eslint/parser": "5.57.0",
         "@swimlane/ngx-charts": "^20.1.2",
         "eslint-plugin-import": "latest",
         "@ngx-translate/core": "^14.0.0",
@@ -61,13 +61,13 @@
         "@types/d3-scale": "^4.0.3",
         "@types/d3-shape": "^3.1.1",
         "autoprefixer": "^10.4.14",
-        "@types/node": "^18.15.5",
-        "tailwindcss": "^3.2.7",
+        "@types/node": "^18.15.11",
+        "tailwindcss": "^3.3.0",
         "typescript": "~4.9.5",
         "zone.js": "~0.13.0",
         "ts-node": "^10.9.1",
         "postcss": "^8.4.21",
-        "eslint": "^8.36.0",
+        "eslint": "^8.37.0",
         "jquery": "^3.6.4",
         "rxjs": "~7.8.0"
     }
