@@ -29,6 +29,7 @@ import { HttpClient, HttpHeaders } from "@angular/common/http";
 import { NotificationService } from "../../services/notifications/notifications.service";
 import { PrivilegesService } from "../../services/privileges.service";
 import { HistoryService } from "../../services/history.service";
+import {PasswordVerificationService} from "../../services/password-verification.service";
 
 @Component({
     selector: 'app-user-profile',
@@ -40,7 +41,6 @@ export class UserProfileComponent implements OnInit {
     loading                     : boolean       = true;
     userId                      : any;
     profile                     : any;
-    errorMessage                : string        = '';
     roles                       : any[]         = [];
     profileForm                 : any[]         = [
         {
@@ -88,12 +88,6 @@ export class UserProfileComponent implements OnInit {
         }
     ];
     disablePasswordModification : boolean       = false;
-    passwordRules               : any           = {
-        minLength: 0,
-        uppercaseMandatory: false,
-        specialCharMandatory: false,
-        numberMandatory: false
-    };
 
     constructor(
         private http: HttpClient,
@@ -106,7 +100,8 @@ export class UserProfileComponent implements OnInit {
         private notify: NotificationService,
         private localeService: LocaleService,
         private historyService: HistoryService,
-        private privilegeService: PrivilegesService
+        private privilegeService: PrivilegesService,
+        private passwordVerification: PasswordVerificationService
     ) { }
 
     ngOnInit() {
@@ -164,20 +159,6 @@ export class UserProfileComponent implements OnInit {
             })
         ).subscribe();
 
-        this.http.get(environment['url'] + '/ws/config/getConfiguration/passwordRules', {headers: this.authService.headers}).pipe(
-            tap((data: any) => {
-                if (data.configuration[0] && data.configuration[0].data.value) {
-                    this.passwordRules = data.configuration[0].data.value;
-                }
-            }),
-            finalize(() => this.loading = false),
-            catchError((err: any) => {
-                console.debug(err);
-                this.notify.handleErrors(err);
-                return of(false);
-            })
-        ).subscribe();
-
         this.http.get(environment['url'] + '/ws/users/getById/' + this.userId, {headers: this.authService.headers}).pipe(
             tap((data: any) => {
                 this.profile = data;
@@ -201,36 +182,17 @@ export class UserProfileComponent implements OnInit {
                 return of(false);
             })
         ).subscribe();
-    }
 
-    checkPasswordValidity(input: any) {
-        if (input.id === 'new_password') {
-            let required: boolean;
-            if (input.control.value) {
-                required = true;
-                if (!input.control.value.match(/[A-Z]/g) && this.passwordRules.uppercaseMandatory) {
-                    this.errorMessage = this.translate.instant('AUTH.password_uppercase_mandatory');
-                } else if (!input.control.value.match(/[0-9]/g) && this.passwordRules.numberMandatory) {
-                    this.errorMessage = this.translate.instant('AUTH.password_number_mandatory');
-                } else if (!input.control.value.match(/[^A-Za-z0-9]/g) && this.passwordRules.specialCharMandatory) {
-                    this.errorMessage = this.translate.instant('AUTH.password_special_char_mandatory');
-                } else if (input.control.value.length < this.passwordRules.minLength && this.passwordRules.minLength !== 0) {
-                    this.errorMessage = this.translate.instant('AUTH.password_min_length', {"min": this.passwordRules.minLength});
-                } else {
-                    this.errorMessage = '';
-                    input.control.setErrors(null);
-                }
-            } else {
-                required = false;
-                this.errorMessage = '';
-                input.control.setErrors(null);
+        this.profileForm.forEach((element: any) => {
+            if (element.id === 'new_password') {
+                element.control.valueChanges.subscribe((value: any) => {
+                    if (value) {
+                        this.passwordVerification.checkPasswordValidityUnique(this.profileForm);
+                    }
+                    this.profileForm.filter((element: any) => element.id === 'old_password')[0].required = !!value;
+                });
             }
-            this.profileForm.forEach(element => {
-                if (element.id === 'old_password') {
-                    element.required = required;
-                }
-            });
-        }
+        });
     }
 
     isValidForm() {
@@ -278,9 +240,8 @@ export class UserProfileComponent implements OnInit {
 
         this.profileForm.forEach(element => {
             if (element.id === field) {
-                if (this.errorMessage !== '' && field === 'new_password') {
-                    element.control.setErrors({});
-                    error = this.errorMessage;
+                if (element.control.errors && element.control.errors.message) {
+                    error = element.control.errors.message;
                 }
                 if (element.required) {
                     error = this.translate.instant('AUTH.field_required');
