@@ -29,6 +29,7 @@ import { catchError, finalize, tap } from "rxjs/operators";
 import { of } from "rxjs";
 import { PrivilegesService } from "../../../../../services/privileges.service";
 import { HistoryService } from "../../../../../services/history.service";
+import {PasswordVerificationService} from "../../../../../services/password-verification.service";
 
 @Component({
     selector: 'app-update',
@@ -43,7 +44,6 @@ export class UpdateUserComponent implements OnInit {
     showPassword                : boolean       = false;
     userId                      : any;
     user                        : any;
-    errorMessage                : string        = '';
     roles                       : any[]         = [];
     userFields                  : any[]         = [
         {
@@ -102,13 +102,6 @@ export class UpdateUserComponent implements OnInit {
     customers                   : any[]         = [];
     userCustomers               : any[]         = [];
     disablePasswordModification : boolean       = false;
-    passwordCurrentlyModified   : boolean       = false;
-    passwordRules               : any           = {
-        minLength: 0,
-        uppercaseMandatory: false,
-        specialCharMandatory: false,
-        numberMandatory: false
-    };
 
     constructor(
         public router: Router,
@@ -121,9 +114,9 @@ export class UpdateUserComponent implements OnInit {
         private notify: NotificationService,
         private historyService: HistoryService,
         public serviceSettings: SettingsService,
-        public privilegesService: PrivilegesService
-    ) {
-    }
+        public privilegesService: PrivilegesService,
+        private passwordVerification: PasswordVerificationService
+    ) {}
 
     ngOnInit(): void {
         this.serviceSettings.init();
@@ -210,19 +203,6 @@ export class UpdateUserComponent implements OnInit {
             })
         ).subscribe();
 
-        this.http.get(environment['url'] + '/ws/config/getConfiguration/passwordRules', {headers: this.authService.headers}).pipe(
-            tap((data: any) => {
-                if (data.configuration[0] && data.configuration[0].data.value) {
-                    this.passwordRules = data.configuration[0].data.value;
-                }
-            }),
-            catchError((err: any) => {
-                console.debug(err);
-                this.notify.handleErrors(err);
-                return of(false);
-            })
-        ).subscribe();
-
         this.http.get(environment['url'] + '/ws/users/getById/' + this.userId, {headers: this.authService.headers}).pipe(
             tap((data: any) => {
                 this.user = data;
@@ -246,6 +226,16 @@ export class UpdateUserComponent implements OnInit {
                 return of(false);
             })
         ).subscribe();
+
+        this.userFields.forEach((element: any) => {
+            if (element.id === 'password_check' || element.id === 'password') {
+                element.control.valueChanges.subscribe((value: any) => {
+                    if (value) {
+                        this.passwordVerification.checkPasswordValidity(this.userFields);
+                    }
+                });
+            }
+        });
     }
 
     isValidForm() {
@@ -258,48 +248,6 @@ export class UpdateUserComponent implements OnInit {
         });
 
         return state;
-    }
-
-    checkPasswordValidity(input: any) {
-        if (input.id === 'password_check' || input.id === 'password') {
-            let required : boolean;
-            if (input.control.value) {
-                required = true;
-                if (!input.control.value.match(/[A-Z]/g) && this.passwordRules.uppercaseMandatory) {
-                    this.errorMessage = this.translate.instant('AUTH.password_uppercase_mandatory');
-                } else if (!input.control.value.match(/[0-9]/g) && this.passwordRules.numberMandatory) {
-                    this.errorMessage = this.translate.instant('AUTH.password_number_mandatory');
-                } else if (!input.control.value.match(/[^A-Za-z0-9]/g) && this.passwordRules.specialCharMandatory) {
-                    this.errorMessage = this.translate.instant('AUTH.password_special_char_mandatory');
-                } else if (input.control.value.length < this.passwordRules.minLength && this.passwordRules.minLength !== 0) {
-                    this.errorMessage = this.translate.instant('AUTH.password_min_length', {"min": this.passwordRules.minLength});
-                } else {
-                    this.errorMessage = '';
-                    this.userFields.forEach(element => {
-                        if (element.id === 'password' && element.control.value !== input.control.value) {
-                            this.errorMessage = this.translate.instant('USER.password_mismatch');
-                        }
-                    });
-                    if (this.errorMessage === '') {
-                        input.control.setErrors(null);
-                    }
-                }
-            } else {
-                this.userFields.forEach(element => {
-                    if (element.id === 'password' && element.control.value === '' && input.control.value === '') {
-                        required = false;
-                        this.errorMessage = '';
-                        input.control.setErrors(null);
-                    }
-                });
-            }
-            this.userFields.forEach(element => {
-                if (element.id === 'password_check' || element.id === 'password') {
-                    this.passwordCurrentlyModified = required;
-                    element.required = required;
-                }
-            });
-        }
     }
 
     onSubmit() {
@@ -330,9 +278,8 @@ export class UpdateUserComponent implements OnInit {
         let error: any;
         this.userFields.forEach(element => {
             if (element.id === field) {
-                if (this.errorMessage !== '' && field === 'password_check') {
-                    element.control.setErrors({});
-                    error = this.errorMessage;
+                if (element.control.errors && element.control.errors.message) {
+                    error = element.control.errors.message;
                 }
                 if (element.required && !(element.value || element.control.value)) {
                     error = this.translate.instant('AUTH.field_required');
@@ -371,10 +318,11 @@ export class UpdateUserComponent implements OnInit {
             cpt = cpt + 1;
         }
 
-        if (!found)
+        if (!found) {
             this.userCustomers.push(customerId);
-        else
+        } else {
             this.userCustomers.splice(cpt, 1);
+        }
 
         this.http.put(environment['url'] + '/ws/users/customers/update/' + this.userId, {'customers': this.userCustomers}, {headers: this.authService.headers},
         ).pipe(
@@ -400,10 +348,11 @@ export class UpdateUserComponent implements OnInit {
             cpt = cpt + 1;
         }
 
-        if (!found)
+        if (!found) {
             this.userForms.push(formId);
-        else
+        } else {
             this.userForms.splice(cpt, 1);
+        }
 
         this.http.put(environment['url'] + '/ws/users/forms/update/' + this.userId, {'forms': this.userForms}, {headers: this.authService.headers},
         ).pipe(
